Break up and refactor WCS_ATT_Display_Cart
